openwebui/ppt/requirements.txt
flask
Flask-Cors
openai
python-pptx
requests